requirements.txt
python-telegram-bot
requests
python-binance
matplotlib
pandas
